requirements.txt
numpy
pywavelets
adi-reader
pandas
matplotlib
statsmodels